+++ App/Sources/Models/Rakah.swift
@@ -15,33 +15,23 @@ typealias Rakat = [Rakah]
 class Rakah {
   // MARK: - Stored Instance Properties
   private let isBeginningOfPrayer: Bool
-  private let includesStandingRecitation: Bool
   private let includesSittingRecitation: Bool
   private let isEndOfPrayer: Bool
-  let standingRecitationFileName: String?
+  let standingRecitationPart: RecitationPart?
+
+  private var includesStandingRecitation: Bool { standingRecitationPart != nil }
 
   // MARK: - Initializer
   init(
     isBeginningOfPrayer: Bool,
-    includesStandingRecitation: Bool,
+    standingRecitationPart: RecitationPart?,
     includesSittingRecitation: Bool,
-    isEndOfPrayer: Bool,
-    excludeStandingRecitationNames: [String] = []
+    isEndOfPrayer: Bool
   ) {
     self.isBeginningOfPrayer = isBeginningOfPrayer
-    self.includesStandingRecitation = includesStandingRecitation
     self.includesSittingRecitation = includesSittingRecitation
     self.isEndOfPrayer = isEndOfPrayer
-
-    if includesStandingRecitation {
-      let allowedStandingRecitationNames = Component.nonOpeningRecitationFileNames.filter {
-        !excludeStandingRecitationNames.contains($0)
-      }
-      self.standingRecitationFileName = allowedStandingRecitationNames.sample
-    }
-    else {
-      self.standingRecitationFileName = nil
-    }
+    self.standingRecitationPart = standingRecitationPart
   }
 
   // MARK: - Instance Methods
@@ -54,11 +44,11 @@ class Rakah {
       components.append(RakahComponent(.taawwudh))
     }
 
-    components.append(RakahComponent(.recitation(fileName: "001_The-Opening")))
+    components.append(RakahComponent(.recitationPart(.init(recitation: .theOpening, partLength: .short))))
 
-    if includesStandingRecitation {
-      let standingRecitation = RakahComponent(.recitation(fileName: standingRecitationFileName!))
-      components.append(standingRecitation)
+    if let standingRecitationPart = standingRecitationPart {
+      let standingRecitationComponent = RakahComponent(.recitationPart(standingRecitationPart))
+      components.append(standingRecitationComponent)
     }
 
     components.append(RakahComponent(.takbir(.bending)))
@@ -100,7 +90,7 @@ extension Rakah {
     case takbir(Position)
     case openingSupplication
     case taawwudh
-    case recitation(fileName: String)
+    case recitationPart(RecitationPart)
     case ruku
     case straighteningUp  // from Ruku
     case sajdah
@@ -108,24 +98,5 @@ extension Rakah {
     case salatulIbrahimiyyah
     case rabbanagh
     case salam(Position)
-
-    static let theOpeningRecitationFileName: String = "001_The-Opening"
-
-    static let nonOpeningRecitationFileNames: [String] = [
-      "103_The-Flight-of-Time", "104_The-Slanderer", "105_The-Elephant", "106_Quraysh", "107_Assistance",
-      "108_Good-in-Abundance", "109_Those-Who-Deny-the-Truth", "110_Succour", "111_The-Twisted-Strands",
-      "112_The-Declaration-of-Gods-Perfection", "113_The-Rising-Dawn", "114_Men",
-    ]
-
-    static var all: [Component] {
-      let nonQuranicComponents: [Component] = [
-        .takbir(.standing), .takbir(.bending), .takbir(.sitting), .takbir(.worship),
-        .openingSupplication, .taawwudh, .ruku, .straighteningUp, .sajdah, .tashahhud,
-        .salatulIbrahimiyyah, .rabbanagh, .salam(.salamRight), .salam(.salamLeft),
-      ]
-      let recitations = ([theOpeningRecitationFileName] + nonOpeningRecitationFileNames)
-        .map { Component.recitation(fileName: $0) }
-      return nonQuranicComponents + recitations
-    }
   }
 }