+++ App/Sources/ScreenFlows/Settings/SettingsFlowController.swift
@@ -21,7 +21,7 @@ class SettingsFlowController: InitialFlowController {
     settingsViewCtrl = SettingsViewController(viewModel: settingsViewModel)
     settingsViewCtrl?.flowDelegate = self
 
-    let navCtrl = BrandedNavigationController(rootViewController: settingsViewCtrl)
+    let navCtrl = UINavigationController(rootViewController: settingsViewCtrl)
     window.rootViewController = navCtrl
 
     if !Defaults.faqClosed {
@@ -47,6 +47,14 @@ extension SettingsFlowController: SettingsFlowDelegate {
     settingsViewModel.showChangingTextName = showChangingTextName
   }
 
+  func setAllowLongerRecitations(_ allowLongerRecitations: Bool) {
+    settingsViewModel.allowLongerRecitations = allowLongerRecitations
+  }
+
+  func setAllowSplittingRecitations(_ allowSplittingRecitations: Bool) {
+    settingsViewModel.allowSplittingRecitations = allowSplittingRecitations
+  }
+
   func showLanguageSettings() {
     UIApplication.shared.open(URL(string: UIApplication.openSettingsURLString)!)
   }
@@ -69,7 +77,11 @@ extension SettingsFlowController: SettingsFlowDelegate {
 
   func startPrayer() {
     guard settingsViewModel.rakatCount > 0 else { return }
-    let prayer = Prayer(rakatCount: UInt(settingsViewModel.rakatCount))
+    let prayer = Prayer(
+      rakatCount: UInt(settingsViewModel.rakatCount),
+      allowLongerRecitations: settingsViewModel.allowLongerRecitations,
+      allowSplittingRecitations: settingsViewModel.allowSplittingRecitations
+    )
 
     let prayerCoordinator = PrayerFlowController(
       prayer: prayer,
@@ -103,7 +115,7 @@ extension SettingsFlowController: SettingsFlowDelegate {
   }
 
   func showFeedbackCommunity() {
-    let communityUrl = URL(string: "https://community.flinesoft.com/c/prayer-app")!
+    let communityUrl = URL(string: "https://links.flinesoft.com/forum/prayer")!
     let safariViewCtrl = SFSafariViewController(url: communityUrl)
 
     settingsViewCtrl.present(safariViewCtrl, animated: true, completion: nil)