+++ App/Sources/ScreenFlows/Settings/SettingsViewModel.swift
@@ -15,37 +15,49 @@ class SettingsViewModel {
 
   // MARK: - Computed Instance Properties
   var rakatCount: Int {
-    get { return Defaults.rakatCount }
+    get { Defaults.rakatCount }
     set { Defaults.rakatCount = newValue }
   }
 
   var fixedTextsSpeedFactor: Double {
-    get { return Defaults.fixedTextsSpeedFactor }
+    get { Defaults.fixedTextsSpeedFactor }
     set { Defaults.fixedTextsSpeedFactor = newValue }
   }
 
   var changingTextSpeedFactor: Double {
-    get { return Defaults.changingTextSpeedFactor }
+    get { Defaults.changingTextSpeedFactor }
     set { Defaults.changingTextSpeedFactor = newValue }
   }
 
   var showChangingTextName: Bool {
-    get { return Defaults.showChangingTextName }
+    get { Defaults.showChangingTextName }
     set { Defaults.showChangingTextName = newValue }
   }
 
+  var allowLongerRecitations: Bool {
+    get { Defaults.allowLongerRecitations }
+    set { Defaults.allowLongerRecitations = newValue }
+  }
+
+  var allowSplittingRecitations: Bool {
+    get { Defaults.allowSplittingRecitations }
+    set { Defaults.allowSplittingRecitations = newValue }
+  }
+
   var movementSoundInstrument: String {
-    get { return Defaults.movementSoundInstrument }
+    get { Defaults.movementSoundInstrument }
     set { Defaults.movementSoundInstrument = newValue }
   }
 }
 
 extension DefaultsKeys {
-  private var defaultInstrument: String { return SettingsViewModel.availableMovementSoundInstruments.first! }
+  private var defaultInstrument: String { SettingsViewModel.availableMovementSoundInstruments.first! }
 
   var rakatCount: DefaultsKey<Int> { .init("RakatCount", defaultValue: 4) }
   var fixedTextsSpeedFactor: DefaultsKey<Double> { .init("FixedTextsSpeedFactor", defaultValue: 1.0) }
   var changingTextSpeedFactor: DefaultsKey<Double> { .init("ChangingTextSpeedFactor", defaultValue: 1.0) }
   var showChangingTextName: DefaultsKey<Bool> { .init("ShowChangingTextName", defaultValue: true) }
+  var allowLongerRecitations: DefaultsKey<Bool> { .init("AllowLongerRecitations", defaultValue: false) }
+  var allowSplittingRecitations: DefaultsKey<Bool> { .init("AllowSplittingRecitations", defaultValue: false) }
   var movementSoundInstrument: DefaultsKey<String> { .init("MovementSoundInstrument", defaultValue: defaultInstrument) }
 }